fix: moved logic into a singleton component to be nested under the redux provider
